+++ src/ChainState.hpp
@@ -26,7 +26,7 @@ namespace ffSCITE {
  *
  * @tparam max_n_genes The maximal number of genes in the dataset.
  */
-template <uint64_t max_n_genes> struct ChainState {
+template <uint32_t max_n_genes> struct ChainState {
 public:
   /**
    * @brief The maximal number of nodes in a mutation tree.
@@ -34,7 +34,7 @@ public:
    * The mutation tree has a node for every gene that may mutate, but also an
    * additional node that represents the totally unmutated state.
    */
-  static constexpr uint64_t max_n_nodes = max_n_genes + 1;
+  static constexpr uint32_t max_n_nodes = max_n_genes + 1;
 
   /**
    * @brief Shorthand for the parent vector type.
@@ -53,7 +53,7 @@ public:
    * @param mutation_tree The mutation tree to use.
    * @param beta The probability of false negatives for the new state.
    */
-  ChainState(ParentVector<max_n_nodes> mutation_tree, double beta)
+  ChainState(ParentVector<max_n_nodes> mutation_tree, float beta)
       : mutation_tree(mutation_tree), beta(beta) {}
 
   /**
@@ -68,8 +68,8 @@ public:
    * @param beta The probability of false negatives for the new state.
    */
   template <typename RNG>
-  static ChainState<max_n_genes> sample_random_state(RNG &rng, uint64_t n_genes,
-                                                     double beta) {
+  static ChainState<max_n_genes> sample_random_state(RNG &rng, uint32_t n_genes,
+                                                     float beta) {
     ParentVectorImpl mutation_tree =
         ParentVectorImpl::sample_random_tree(rng, n_genes + 1);
     return ChainState<max_n_genes>(mutation_tree, beta);
@@ -94,6 +94,6 @@ public:
   /**
    * @brief The (proposed) beta error rate.
    */
-  double beta;
+  float beta;
 };
 } // namespace ffSCITE
