+++ src/Parameters.hpp
@@ -333,9 +333,9 @@ public:
    * For every chain, a new random starting state is initialized and @ref
    * get_chain_length steps are executed from this state.
    *
-   * @return uint64_t The number of chains to execute.
+   * @return uint32_t The number of chains to execute.
    */
-  uint64_t get_n_chains() const { return n_chains.value(); }
+  uint32_t get_n_chains() const { return n_chains.value(); }
 
   /**
    * @brief Set the number of chains to execute.
@@ -345,7 +345,7 @@ public:
    *
    * @param n_chains The new number of chains to execute.
    */
-  void set_n_chains(uint64_t n_chains) { this->n_chains = n_chains; }
+  void set_n_chains(uint32_t n_chains) { this->n_chains = n_chains; }
 
   /**
    * @brief Get the number of steps to execute on every chain.
@@ -354,9 +354,9 @@ public:
    * state, computing its likelihood and deciding whether the new state should
    * be the new state of the chain.
    *
-   * @return uint64_t The number of steps to execute on every chain.
+   * @return uint32_t The number of steps to execute on every chain.
    */
-  uint64_t get_chain_length() const { return chain_length.value(); }
+  uint32_t get_chain_length() const { return chain_length.value(); }
 
   /**
    * @brief Set the number of steps to execute on every chain.
@@ -367,7 +367,7 @@ public:
    *
    * @param chain_length The new number of steps to execute on every chain.
    */
-  void set_chain_length(uint64_t chain_length) {
+  void set_chain_length(uint32_t chain_length) {
     this->chain_length = chain_length;
   }
 
@@ -378,9 +378,9 @@ public:
    * can not be reallocated. Therefore, an upper bound for the maximal number of
    * optimal states is needed.
    *
-   * @return uint64_t The maximal number of optimal states to return.
+   * @return uint32_t The maximal number of optimal states to return.
    */
-  uint64_t get_max_n_best_states() const { return max_n_best_states; }
+  uint32_t get_max_n_best_states() const { return max_n_best_states; }
 
   /**
    * @brief Set the maximal number of optimal states to return.
@@ -392,17 +392,17 @@ public:
    * @param max_n_best_states The new maximal number of optimal states to
    * return.
    */
-  void set_max_n_best_states(uint64_t max_n_best_states) {
+  void set_max_n_best_states(uint32_t max_n_best_states) {
     this->max_n_best_states = max_n_best_states;
   }
 
   /**
    * @brief Get the mean value of the probability for false positives (alpha).
    *
-   * @return double The mean value of the probability for false positives
+   * @return float The mean value of the probability for false positives
    * (alpha).
    */
-  double get_alpha_mean() const { return alpha_mean; }
+  float get_alpha_mean() const { return alpha_mean; }
 
   /**
    * @brief Set the mean value of the probability for false positives (alpha).
@@ -410,15 +410,15 @@ public:
    * @param alpha_mean The new mean value of the probability for false positives
    * (alpha).
    */
-  void set_alpha_mean(double alpha_mean) { this->alpha_mean = alpha_mean; }
+  void set_alpha_mean(float alpha_mean) { this->alpha_mean = alpha_mean; }
 
   /**
    * @brief Get the mean value for the probability for false negatives (beta).
    *
-   * @return double The mean value for the probability for false negatives
+   * @return float The mean value for the probability for false negatives
    * (beta).
    */
-  double get_beta_mean() const { return beta_mean; }
+  float get_beta_mean() const { return beta_mean; }
 
   /**
    * @brief Set the mean value for the probability for false negatives (beta).
@@ -426,16 +426,16 @@ public:
    * @param beta_mean The new mean value for the probability for false negatives
    * (beta).
    */
-  void set_beta_mean(double beta_mean) { this->beta_mean = beta_mean; }
+  void set_beta_mean(float beta_mean) { this->beta_mean = beta_mean; }
 
   /**
    * @brief Get the standard derivation for the probability for false negatives
    * (beta).
    *
-   * @return double The standard derivation for the probability for false
+   * @return float The standard derivation for the probability for false
    * negatives (beta).
    */
-  double get_beta_sd() const { return beta_sd; }
+  float get_beta_sd() const { return beta_sd; }
 
   /**
    * @brief Set the standard derivation for the probability for false negatives
@@ -444,17 +444,17 @@ public:
    * @param beta_sd The new standard derivation for the probability for false
    * negatives (beta).
    */
-  void set_beta_sd(double beta_sd) { this->beta_sd = beta_sd; }
+  void set_beta_sd(float beta_sd) { this->beta_sd = beta_sd; }
 
   /**
    * @brief Get the probability that the beta value is modified in a chain step.
    *
    * The returned value of this method is bound to the interval [0,1].
    *
-   * @return double The probability that the beta value is modified in a chain
+   * @return float The probability that the beta value is modified in a chain
    * step.
    */
-  double get_prob_beta_change() const {
+  float get_prob_beta_change() const {
     return prob_beta_change / get_sum_of_move_probs();
   }
 
@@ -467,7 +467,7 @@ public:
    * @param prob_beta_change The new probability that the beta value is modified
    * in a chain step.
    */
-  void set_prob_beta_change(double prob_beta_change) {
+  void set_prob_beta_change(float prob_beta_change) {
     if (prob_beta_change < 0.0) {
       throw std::out_of_range("Move probabilities may not be negative.");
     }
@@ -480,10 +480,10 @@ public:
    *
    * The returned value of this method is bound to the interval [0,1].
    *
-   * @return double The probability that a node and its subtree is moved to
+   * @return float The probability that a node and its subtree is moved to
    * another node in a chain step.
    */
-  double get_prob_prune_n_reattach() const {
+  float get_prob_prune_n_reattach() const {
     return prob_prune_n_reattach / get_sum_of_move_probs();
   }
 
@@ -497,7 +497,7 @@ public:
    * @param prob_prune_n_reattach The new probability that a node and its
    * subtree is moved to another node in a chain step.
    */
-  void set_prob_prune_n_reattach(double prob_prune_n_reattach) {
+  void set_prob_prune_n_reattach(float prob_prune_n_reattach) {
     if (prob_prune_n_reattach < 0.0) {
       throw std::out_of_range("Move probabilities may not be negative.");
     }
@@ -510,10 +510,10 @@ public:
    *
    * The returned value of this method is bound to the interval [0,1].
    *
-   * @return double The probability that two node labels are swapped in a chain
+   * @return float The probability that two node labels are swapped in a chain
    * step.
    */
-  double get_prob_swap_nodes() const {
+  float get_prob_swap_nodes() const {
     return prob_swap_nodes / get_sum_of_move_probs();
   }
 
@@ -527,7 +527,7 @@ public:
    * @param prob_swap_nodes The new probability that two node labels are swapped
    * in a chain step.
    */
-  void set_prob_swap_nodes(double prob_swap_nodes) {
+  void set_prob_swap_nodes(float prob_swap_nodes) {
     if (prob_swap_nodes < 0.0) {
       throw std::out_of_range("Move probabilities may not be negative.");
     }
@@ -539,10 +539,10 @@ public:
    *
    * The returned value of this method is bound to the interval [0,1].
    *
-   * @return double The probability that two subtrees are swapped in a chain
+   * @return float The probability that two subtrees are swapped in a chain
    * step.
    */
-  double get_prob_swap_subtrees() const {
+  float get_prob_swap_subtrees() const {
     return prob_swap_subtrees / get_sum_of_move_probs();
   }
 
@@ -555,7 +555,7 @@ public:
    * @param prob_swap_subtrees The new probability that two subtrees are swapped
    * in a chain step.
    */
-  void set_prob_swap_subtrees(double prob_swap_subtrees) {
+  void set_prob_swap_subtrees(float prob_swap_subtrees) {
     if (prob_swap_subtrees < 0.0) {
       throw std::out_of_range("Move probabilities may not be negative.");
     }
@@ -565,10 +565,10 @@ public:
   /**
    * @brief Get the scaling factor to the standard derivation for beta changes.
    *
-   * @return double The scaling factor to the standard derivation for beta
+   * @return float The scaling factor to the standard derivation for beta
    * changes.
    */
-  double get_beta_jump_scaling_chi() const { return beta_jump_scaling_chi; }
+  float get_beta_jump_scaling_chi() const { return beta_jump_scaling_chi; }
 
   /**
    * @brief Set the scaling factor to the standard derivation for beta changes.
@@ -576,7 +576,7 @@ public:
    * @param beta_jump_scaling_chi The new scaling factor to the standard
    * derivation for beta changes.
    */
-  void set_beta_jump_scaling_chi(double beta_jump_scaling_chi) {
+  void set_beta_jump_scaling_chi(float beta_jump_scaling_chi) {
     this->beta_jump_scaling_chi = beta_jump_scaling_chi;
   }
 
@@ -587,9 +587,9 @@ public:
    * gamma > 1.0, the chain focuses on local exploration and for gamma < 1.0,
    * the chain focuses on global exploration.
    *
-   * @return double The gamma factor for the exploration behavior.
+   * @return float The gamma factor for the exploration behavior.
    */
-  double get_gamma() const { return gamma; }
+  float get_gamma() const { return gamma; }
 
   /**
    * @brief Set the gamma factor for the exploration behavior.
@@ -600,7 +600,7 @@ public:
    *
    * @param gamma The new gamma factor for the exploration behavior.
    */
-  void set_gamma(double gamma) { this->gamma = gamma; }
+  void set_gamma(float gamma) { this->gamma = gamma; }
 
   /**
    * @brief Get the seed for the URNGs.
@@ -627,21 +627,21 @@ public:
   void set_seed(uint32_t seed) { this->seed = seed; }
 
 private:
-  double get_sum_of_move_probs() const {
+  float get_sum_of_move_probs() const {
     return prob_beta_change + prob_prune_n_reattach + prob_swap_nodes +
            prob_swap_subtrees;
   }
 
   std::optional<std::string> input_path, output_path_base;
 
-  std::optional<uint64_t> n_chains, chain_length;
-  uint64_t max_n_best_states;
+  std::optional<uint32_t> n_chains, chain_length;
+  uint32_t max_n_best_states;
 
-  double alpha_mean, beta_mean, beta_sd;
-  double prob_beta_change, prob_prune_n_reattach, prob_swap_nodes,
+  float alpha_mean, beta_mean, beta_sd;
+  float prob_beta_change, prob_prune_n_reattach, prob_swap_nodes,
       prob_swap_subtrees;
-  double beta_jump_scaling_chi;
-  double gamma;
+  float beta_jump_scaling_chi;
+  float gamma;
 
   std::optional<uint32_t> seed;
 };
